feat: refactors creds
